pyproject.toml
[tool.poetry]
name = "FastQL"
version = "1.0.0"
description = "A FullStack Template to start a FastAPI Project Based On GraphQL and pydantic"
authors = ["Yasser Tahiri"]

[tool.poetry.dependencies]
aio-pika = "^6.8.0"
aiohttp = "^3.8.1"
aiormq = "^3.3.1"
aiosignal = "^1.2.0"
alembic = "^1.7.5"
anyio = "^3.4.0"
ariadne = "^0.14.0"
asgiref = "^3.4.1"
async-exit-stack = "^1.0.1"
async-generator = "^1.10"
async-timeout = "^4.0.1"
asyncpg = "^0.25.0"
attrs = "^21.2.0"
bcrypt = "^3.2.0"
black = "^21.11b1"
certifi = "^2021.10.8"
cffi = "^1.15.0"
cfgv = "^3.3.1"
charset-normalizer = "^2.0.8"
click = "^8.0.3"
databases = "^0.5.3"
debugpy = "^1.5.1"
distlib = "^0.3.3"
ecdsa = "^0.17.0"
fastapi = "^0.70.0"
filelock = "^3.4.0"
frozenlist = "^1.2.0"
graphql-core = "^3.1.6"
greenlet = "^1.1.2"
h11 = "^0.12.0"
httpcore = "^0.14.3"
httpx = "^0.21.1"
identify = "^2.4.0"
idna = "^3.3"
iniconfig = "^1.1.1"
Mako = "^1.1.6"
MarkupSafe = "^2.0.1"
multidict = "^5.2.0"
mypy-extensions = "^0.4.3"
nodeenv = "^1.6.0"
packaging = "^21.3"
pamqp = "^2.3.0"
passlib = "^1.7.4"
pathspec = "^0.9.0"
platformdirs = "^2.4.0"
pluggy = "^1.0.0"
pre-commit = "^2.15.0"
psycopg2-binary = "^2.9.2"
py = "^1.11.0"
pyasn1 = "^0.4.8"
pycparser = "^2.21"
pydantic = "^1.8.2"
pyparsing = "^3.0.6"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
python-decouple = "^3.5"
python-graphql-client = "^0.4.3"
python-jose = "^3.3.0"
pywsitest = "^0.3.8"
PyYAML = "^6.0"
regex = "^2021.11.10"
requests = "^2.26.0"
rfc3986 = "^1.5.0"
rsa = "^4.8"
six = "^1.16.0"
sniffio = "^1.2.0"
SQLAlchemy = "^1.4.27"
starlette = "^0.16.0"
toml = "^0.10.2"
tomli = "^1.2.2"
typing_extensions = "^4.0.0"
urllib3 = "^1.26.7"
uvicorn = "^0.15.0"
virtualenv = "^20.10.0"
websockets = "^10.1"
yarl = "^1.7.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
